requirements.txt
ansicon==1.89.0
APScheduler==3.6.3
arrow==1.2.2
asgiref==3.5.1
astroid==2.12.9
blessed==1.19.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.5
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.5.1
Django==4.0.4
django-cors-headers==3.13.0
django-crum==0.7.9
django-environ==0.8.1
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-picklefield==3.0.1
django-q==1.3.9
django-q-email==5.0.1
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-seed==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.8.0
Faker==13.12.0
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
jsonfield==3.1.0
lazy-object-proxy==1.7.1
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
openapi-codec==1.3.2
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-telegram-bot==13.12
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
swapper==1.3.0
tomli==2.0.1
tomlkit==0.11.4
toposort==1.7
tornado==6.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
